Start of using state machine for device state
This just creates a first draft state interface and a quick dummy state
to use as an example. Will likely need to pass more to the interface
APIs before this is fully ready and may want to add a state manager
later that owns the raw state pointer and handles transitions.