Move QD state to new states files
Make QD_STATE an enum instead of a simple byte so values can be enforced
and move it to states.hpp and states.cpp so we can later access it from
files other than main. This will get refactored again later in the state
machine change, but is needed for now to support other refactor work.