Add basic P2P wireless capabilities and a remote player discovery (#36)
* Copy files from initial-esp-now branch

* Finish esp integration with newer main branch

This branch replaces the old initial-esp-now branch and this commit
completes what should be the last of the needed integration. This builds
but is not at all tested yet.

* Switch over to using ESP_LOG for logging

This does not require CDC on boot and also adds extra file information
to messages, plus adds tagging and log levels, so is all around a
better solution. You can now see remote player discovery messages
when using Upload and Monitor command in VSCode.

---------

Co-authored-by: Daniel Lazzari Jr <neoskye@gmail.com>